moved STL datatype traits to contrib/stl
Update #548
